+++ addons/base_address_extended/models/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import base_address_extended
+from . import res_company
+from . import res_country
+from . import res_partner
